+++ app/src/main/java/com/robinkuiper/cardsscorekeeper/data/game/boerenBridge/rounds/PredictedRound.java
@@ -5,15 +5,15 @@ import java.util.Map;
 public class PredictedRound {
     private final RoundData roundData;
 
-    public PredictedRound(int playerCount, int roundCount, Map<Integer, Integer> predictions) {
+    public PredictedRound(int playerCount, int roundCount, Map<Long, Integer> predictions) {
         this.roundData = new RoundData(playerCount, roundCount, predictions);
     }
 
-    public FinishedRound addScores(Map<Integer, Integer> scores) {
+    public FinishedRound addScores(Map<Long, Integer> scores) {
         return roundData.addScores(scores);
     }
 
-    public Map<Integer, Integer> getPredictions() {
+    public Map<Long, Integer> getPredictions() {
         return roundData.getPredictions();
     }
 }
